requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
Faker==39.0.0
fastapi==0.121.1
Flask==3.1.2
fpdf==1.7.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.5
reportlab==4.4.7
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.4
